+++ src/main/java/roomescape/reservation/controller/MemberReservationController.java
@@ -5,8 +5,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.auth.dto.LoginMember;
 import roomescape.reservation.dto.MemberReservationCreateRequest;
-import roomescape.reservation.dto.MyReservationResponse;
 import roomescape.reservation.dto.MemberReservationResponse;
+import roomescape.reservation.dto.MyReservationResponse;
 import roomescape.reservation.service.ReservationService;
 
 import java.util.List;
